refactor: 바텀 시트 내용물을 순수 뷰 컴포넌트로 분리
